reorganize readback code
