+++ src/term/mod.rs
@@ -13,14 +13,12 @@ pub mod builtins;
 pub mod check;
 pub mod display;
 pub mod load_book;
-pub mod net_to_term;
 pub mod parser;
 pub mod readback;
 pub mod term_to_net;
 pub mod transform;
 
 pub use hvmc::ops::{IntOp, Op, Ty as OpType};
-pub use net_to_term::{net_to_term, ReadbackError};
 pub use readback::readback;
 pub use term_to_net::{book_to_nets, term_to_net};
