+++ tests/golden_tests.rs
@@ -1,10 +1,11 @@
 use hvml::{
   compile_book, desugar_book,
   diagnostics::{Diagnostics, DiagnosticsConfig, Severity, ToStringVerbose},
-  net::hvmc_to_net::hvmc_to_net,
   run_book,
   term::{
-    load_book::do_parse_book, net_to_term::net_to_term, parser::TermParser, term_to_net, term_to_net::Labels,
+    load_book::do_parse_book,
+    parser::TermParser,
+    term_to_net::{term_to_net, Labels},
     AdtEncoding, Book, Ctx, Name,
   },
   CompileOpts, RunOpts,
@@ -201,11 +202,10 @@ fn readback() {
   run_golden_test_dir(function_name!(), &|code, _| {
     let net = hvmc::ast::Net::from_str(code)?;
     let book = Book::default();
-    let compat_net = hvmc_to_net(&net);
     let mut non_linear_diags = Diagnostics::default();
-    let non_linear_term = net_to_term(&compat_net, &book, &Labels::default(), false, &mut non_linear_diags);
+    let non_linear_term = hvml::term::readback(&net, &book, &Labels::default(), false, &mut non_linear_diags);
     let mut linear_diags = Diagnostics::default();
-    let linear_term = net_to_term(&compat_net, &book, &Labels::default(), true, &mut linear_diags);
+    let linear_term = hvml::term::readback(&net, &book, &Labels::default(), true, &mut linear_diags);
     Ok(format!(
       "non-linear:\n{}{}\n\nlinear:\n{}{}",
       non_linear_diags, non_linear_term, linear_diags, linear_term
